QUrl::FormattingOptions flag arguments didn't work for Qt 5
QUrl::FormattingOptions has the peculiar situation that it actually consists of values from QUrl::UrlFormattingOption and QUrl::ComponentFormattingOption.
It seems the moc doesn't support this situation, so at least support the UrlFormattingOption values.
(I left the old entry in to support Qt 4, but I don't know if PythonQt even still works with Qt 5)