Fix a C++20 compatibility issue with an unnamed struct (P1766R1)
As a result of P1766R1 [1], `PyHeapTypeObject` can no longer be added
to an unnamed `typedef struct`. On Visual Studio 2019 ≥16.6, this
causes the warning C5208 in C++14 and C++17 mode, and the error C7626
with `/std:c++latest` or the future `/std:c++20`. Giving a name to the
unnamed struct is the simplest fix here [2].

[1]: http://wg21.link/p1766r1
[2]: https://docs.microsoft.com/en-us/cpp/error-messages/compiler-warnings/c5208