Fix handling of compiler standard under qmake6
Qt6 *requires* support for the C++17 standard.  This is the default with
current versions of g++ (GNU) and is also the default standard level
enabled by qmake6 (see https://doc.qt.io/qt-6/qmake-variable-reference.html)
however QMAKE_CXXFLAGS is set to force -std=c++98 (-ansi is an alias for
this).  The qmake CONFIG options try to explicitly enable c++11 (not
sufficient for Qt6) and this does result in qmake5 setting -std=c++11.
This does not happen, or does not happen always, with qmake6 and,
anyway, is insufficient.

It is also the case that a standard to use should not be set to
conflicing values in two different places.  Ideally it should be left to
the person who supplies the compiler.  (cmake does, in fact, provide
flags to ensure a suitable compiler is used.)

This fix simply removes both conflicting attempts to set the C++
standard.  It therefore allows compilation with later versions of the
standard (e.g. c++20).  It works with both Qt5 and Qt6.  Tested with:

qmake5:
  Make version 3.1
  Using Qt version 5.15.10 in /usr/lib64

qmake6:
  QMake version 3.1
  Using Qt version 6.5.3 in /usr/lib64

With g++ and glibc 2.38:
  g++ (Gentoo 13.2.1_p20230826 p7) 13.2.1 20230826

Signed-off-by: John Bowler <jbowler@acm.org>