Qt6: QList<T>operator== requires const T::operator==
This causes a g++ compile error in codemodel.cpp (the error message is
extremely confusing).  It "works" in Qt5 but adding the "const" in the
relevant places does not break the compile.  Therefore:

bool TypeInfo::operator==(const TypeInfo&)

does not modify the LHS; this means that adding the const is safe (there
would be a compile error if somoething in the implementation did modify
part of the LHS).

It makes sense to me that Qt6 would do something to prevent an ==
operator with a mutable LHS.  So far as I can see QList would have to
make a complete copy of the LHS under some circumstances resulting in a
potentially significant performance hit (depending on the complexity of
the type involved).

This is a harmless but possibly beneficial change to both 'master' and
'qt6' branches.

Signed-off-by: John Bowler <jbowler@acm.org>