Qt6: QList<T>operator== requires const T::operator==
This causes a g++ compile error in codemodel.cpp (see the end of this
comment).  Apparently it "works" in Qt5 but adding the "const" in the
relevant places does not break the compile, so:

bool TypeInfo::operator==(const TypeInfo&)

Apparently does not modify the LHS; this means that adding the const is
safe (there would be a compile error if somoething in the implementation
did modify part of the LHS).

It makes sense to me that Qt6 would do something to prevent an ==
operator with a mutable LHS.  So far as I can see QList would have to
make a complete copy of the LHS under some circumstances resulting in a
potentially significant performance hit (depending on the complexity of
the type involved).

This is a harmless but possibly beneficial change to both 'master' and 'qt6' branches.

Here is the precise compiler message from g++ 13.2.1:

parser/codemodel.cpp: In member function ‘bool TypeInfo::operator==(const TypeInfo&)’:
parser/codemodel.cpp:237:51: error: no match for ‘operator==’ (operand types are ‘QList<TypeInfo>’ and ‘const QList<TypeInfo>’)
  237 |     && (!m_flags.m_functionPointer || m_arguments == other.m_arguments);
      |                                       ~~~~~~~~~~~ ^~ ~~~~~~~~~~~~~~~~~
      |                                       |                    |
      |                                       QList<[...]>         QList<[...]>
In file included from /usr/include/qt6/QtCore/QList:1,
                 from parser/codemodel_fwd.h:47,
                 from parser/codemodel.h:46:
/usr/include/qt6/QtCore/qlist.h:322:56: note: candidate: ‘template<class U> QTypeTraits::compare_eq_result_container<QList<T>, U> QList<T>::operator==(const QList<T>&) const [with T = TypeInfo]’
  322 |     QTypeTraits::compare_eq_result_container<QList, U> operator==(const QList &other) const
      |                                                        ^~~~~~~~
/usr/include/qt6/QtCore/qlist.h:322:56: note:   template argument deduction/substitution failed:
In file included from /usr/include/qt6/QtCore/qglobal.h:13,
                 from /usr/include/qt6/QtCore/qshareddata.h:7,
                 from /usr/include/qt6/QtCore/QSharedData:1,
                 from parser/codemodel_pointer.h:45,
                 from parser/codemodel_fwd.h:46:
/usr/lib/gcc/x86_64-pc-linux-gnu/13/include/g++-v13/type_traits: In substitution of ‘template<bool _Cond, class _Tp> using std::enable_if_t = typename std::enable_if::type [with bool _Cond = false; _Tp = bool]’:
/usr/include/qt6/QtCore/qtypeinfo.h:319:7:   required by substitution of ‘template<class Container, class ... T> using QTypeTraits::compare_eq_result_container = std::enable_if_t<conjunction_v<std::disjunction<std::is_base_of<Container, T>, QTypeTraits::has_operator_equal<T> >...>, bool> [with Container = QList<TypeInfo>; T = {TypeInfo}]’
/usr/include/qt6/QtCore/qlist.h:322:56:   required by substitution of ‘template<class U> QTypeTraits::compare_eq_result_container<QList<TypeInfo>, U> QList<TypeInfo>::operator==(const QList<TypeInfo>&) const [with U = TypeInfo]’
parser/codemodel.cpp:237:60:   required from here
/usr/lib/gcc/x86_64-pc-linux-gnu/13/include/g++-v13/type_traits:2610:11: error: no type named ‘type’ in ‘struct std::enable_if<false, bool>’
 2610 |     using enable_if_t = typename enable_if<_Cond, _Tp>::type;
      |           ^~~~~~~~~~~
make[1]: *** [Makefile:1112: codemodel.o] Error 1

Signed-off-by: John Bowler <jbowler@acm.org>