Eliminate warnings about anonymous structs
These are not valid ISO-C and therefore -Wpendantic (-pedantic) causes
warnings to be emitted.  In practice the code can either fail
compilation (as it would with -Werror) or work correctly so turning
these warnings off is harmless and removes confusing and misleading
reports.

A better fix is to give the structs a name (within the containing union)
and change all the references to include the name.  This is trivial but
changes a lot of lines so I'm not happy doing that because of the likely
merge conflicts.

Signed-off-by: John Bowler <jbowler@acm.org>