Qt6: QTextStream compatibility
This provides compatibility with the removal of the global ::endl
for use with QTextStream; Qt6 requires Qt::endl.

It also includes Qt6 specific changes necessary to deal with the
replacement of QTextCodec by QStringConverter and the slight changes
to the default setup of QTextStream (to UTF-8 with auto-detect of
UTF-16).

Signed-off-by: John Bowler <jbowler@acm.org>