generated_cpp_511 from Qt5.11.3
These files are generated by the a current build of pythonqt_generator
from the qt6 branch; source code identical to commit:

38979b78a324494f2603ae009922dc13e776069f

Note that the actual files were generated before that commit was
generated from what appears to be a rebase against master.  The source
files used are identical to those in the commit.

Signed-off-by: John Bowler <jbowler@acm.org>