Hopefully make generator immune against iterator order changes for typeMap
Previously the includes and generated methods for e.g. QItemSelection would change depending on the order in which the AbstractMetaClass entries were generated, which in turn would change which definition of QList would be seen first.
I now try to mark the "main" definition containing the actual methods and such.
For Qt 5.15 this seems to work at least, but I admit that I still don't understand the generator completely.