Introduce alias names for types
Add an entry "aliases" to a type entry in the type database, these types
will automatically be changed to the main type in the generated code.

This fixes the problem that in Qt 6.7 QAudio enums are used under the
QtAudio alias. These enum names are now automatically mapped to the
QAudio name with the right alias entries in the type database.

This doesn't make QAudio available as QtAudio in Python, though.